\name{d1}
\alias{d1}
\docType{data}
\title{
Example data for the SETPath method - dataset from class 1
}
\description{
An example dataset from one class, with 25 samples and 80 genes.}
\usage{data(setpath.data)}
\format{
  The format is:
d1          : num [1:25, 1:80] 0.701 -2.107 0.624 -0.885 0.344 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:25] "control1" "control2" "control3" "control4" ...
  .. ..$ : chr [1:80] "gene1" "gene2" "gene3" "gene4" ...
}

\keyword{datasets}