\name{d2}
\alias{d2}
\docType{data}
\title{
Example data for the SETPath method - dataset from class 2
}
\description{
An example dataset from one class, with 50 samples and 80 genes.}
\usage{data(setpath.data)}
\format{
  The format is:
d2          : num [1:50, 1:80] 1.003 -1.323 0.906 0.708 1.131 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:50] "case1" "case2" "case3" "case4" ...
  .. ..$ : chr [1:80] "gene1" "gene2" "gene3" "gene4" ...
}

\keyword{datasets}