\name{pathwaygenes}
\alias{pathwaygenes}
\docType{data}
\title{
Defines the gene memberships of the pathways in the example dataset
}
\description{
A list in which each entry is a character string giving the names of the genes belonging to a pathway}
\usage{data(setpath.data)}
\format{
  The format is:
List of 4
  ..$ : chr [1:20] "gene63" "gene33" "gene21" "gene19" ...
  ..$ : chr [1:50] "gene19" "gene25" "gene22" "gene30" ...
  ..$ : chr [1:25] "gene72" "gene32" "gene62" "gene76" ...
  ..$ : chr [1:30] "gene42" "gene7" "gene45" "gene39" ...
}

\keyword{datasets}

