\name{pathwaynames}
\alias{pathwaynames}
\docType{data}
\title{
Names of the pathways used in the example.
}
\description{
A character vector giving the names of each pathway in the example.}
\usage{data(setpath.data)}
\format{
  The format is:
chr [1:4] "pathway 1" "pathway 2" "pathway 3" "pathway 4"
}

\keyword{datasets}

