\name{setpath.data}
\alias{setpath.data}
\docType{data}
\title{
Example data for the SETPath method
}
\description{
An example dataset containing genes from 4 pathways.}
\usage{data(setpath.data)}
\format{
  The format is:
List of 4
 $ d1          : num [1:25, 1:80] 0.701 -2.107 0.624 -0.885 0.344 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:25] "control1" "control2" "control3" "control4" ...
  .. ..$ : chr [1:80] "gene1" "gene2" "gene3" "gene4" ...
 $ d2          : num [1:50, 1:80] 1.003 -1.323 0.906 0.708 1.131 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:50] "case1" "case2" "case3" "case4" ...
  .. ..$ : chr [1:80] "gene1" "gene2" "gene3" "gene4" ...
 $ pathwaynames: chr [1:4] "pathway 1" "pathway 2" "pathway 3" "pathway 4"
 $ pathwaygenes:List of 4
  ..$ : chr [1:20] "gene63" "gene33" "gene21" "gene19" ...
  ..$ : chr [1:50] "gene19" "gene25" "gene22" "gene30" ...
  ..$ : chr [1:25] "gene72" "gene32" "gene62" "gene76" ...
  ..$ : chr [1:30] "gene42" "gene7" "gene45" "gene39" ...
}

\keyword{datasets}

